requirements.txt
streamlit
chromadb
langchain
langchain-openai
langchain-google-genai
python-dotenv
pandas
pypdf
langchain-experimental
sentence-transformers
langchain-huggingface
